+++ src/atoms/mainContent.ts
@@ -1,14 +1,15 @@
 import { RefObject } from "react";
 import { atom } from "recoil";
+import { MainContent, MainScrollRef } from "../types";
 
 const initContent = null;
-export type MainContent = number | null;
+
 export const mainContent = atom<MainContent>({
   key: "mainContent",
   default: initContent,
 });
 const initRef = null;
-export type MainScrollRef = HTMLElement | null;
+
 export const mainScrollRef = atom<MainScrollRef>({
   key: "mainScrollRef",
   default: initRef,
