+++ src/components/Calendar.tsx
@@ -17,15 +17,10 @@ import { useState } from "react";
 import { useRecoilState } from "recoil";
 import { dateSelected } from "../atoms/date";
 import icons from "../components/icons";
+import { CalendarProps } from "../types";
 import { cls } from "../utils";
 
-interface Props {
-  checkedDates?: Date[];
-  className?: string;
-  selectable?: boolean;
-  selectedDate?: Date;
-}
-export const CalenderDrawer = () => {
+export const CalendarDrawer = () => {
   return (
     <icons.Calendar
       className="w-16 transition-colors py-2 pr-2 h-14 flex justify-end items-center rounded-r-xl bg-primary-200 text-gray-900 group hover:bg-primary-600 hover:text-gray-100"
@@ -38,7 +33,7 @@ const Calendar = ({
   className,
   selectable,
   selectedDate,
-}: Props) => {
+}: CalendarProps) => {
   let today = startOfToday();
   let [dateChosen, setdateChosen] = useState(today);
   let [currentMonth, setCurrentMonth] = useState(format(today, "MMM-yyyy"));
@@ -63,7 +58,6 @@ const Calendar = ({
   const [year, month] = format(firstDayCurrentMonth, "yyyy MMMM", {
     locale: ko,
   }).split(" ");
-  useEffect(() => {}, []);
   return (
     <div className={cls(className)}>
       <div className="flex items-center border-b-[4px] border-dotted pb-3">
